Ensure appsettings.json is always copied to output
The project file `HealthMed.API.AgendamentoConsulta.csproj` has been updated to include a new `<ItemGroup>` element. This element specifies the `appsettings.json` file with a `<Content>` element and sets its `CopyToOutputDirectory` property to `Always`. This change ensures that the `appsettings.json` file is always copied to the output directory during the build process.